pkg: Move proposal logic into separate package
